requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pyyaml
ruamel.yaml
python-multipart
itsdangerous
jinja2